first try
didn't  come out qs good as I'd hoped. the concept is much better in my mind.
this method of objectifying pure-desire also doesn't convey the symbolic tragedy to it, cuase it just looks like a  stiientifc principle. it's only when these events become symbolic, do they take on the form of tragedy